requirements.txt
Django>=4.0
djangorestframework==3.14.0
pygments==2.17.2
django-cors-headers==4.3.1
django-filter==23.5
requests==2.32.2
celery[redis]
django-celery-beat==2.5.0 
yt-dlp==2024.7.16
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0